+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Fixed
+ - Prevent HomeKit's default `Auto` target value from changing an IR thermostat to Auto when an Apple Home automation only turns it on.
+
 ## [2.6.6] - (2026.8.3)
 
 ### Fixed